Client side implementation.
Change-Id: I4ff1c3148970edb8820a30dfda607a8aff3f7a52

incomplete server side impl.

Change-Id: I3a52c931a4bd60a747ae47737f78df23046d1053

Passed TestHSync

Change-Id: Ia70db6af5ed295fdb8fdc55b998a8233dc86c416

Fix bug.

When write finishes a full chunk, the chunk is written to DataNode, but the currentChuk is not updated.
Therefore, the flusher need to check if the chunk is full or not, and if full, skip writing chunk (because it is already written out).

Change-Id: I8685b1894466ec7ebd1bc80c49809523dac4855e

Remove debug messages.

Change-Id: I6678899c7f03317ad5fbd02298e80dd534a3308a