HDDS-13356. Replace goofys with mountpoint-s3
This commit replaces the default mount command from 'goofys' to 'mountpoint-s3' in CsiServer.java. This change aligns with the ongoing efforts to support mountpoint-s3 as the preferred S3 client for CSI volumes.

Change-Id: Icb1a82c8afd242a265ec79baa2e3a3cef2bd8a88

HDDS-13356. Correct mountpoint-s3 command parameters.

Corrected the default mount command for mountpoint-s3 in CsiServer.java to ensure the bucket name and mount path are passed as positional arguments, and the endpoint URL is passed with the --endpoint-url flag, as per mountpoint-s3 documentation.

Change-Id: I4736134a0c93935f8bb16668c34031bab19bf1dc

Configure --force-path-style

Change-Id: I26c8fda91afd956e7cdecff1cdd63b925268b5fd